requirements.txt
Django==4.1
django-treebeard==4.5.1
django-seed==0.3.1
django-filter==22.1
DateTime==4.5
Faker==14.1.0
typing==3.7.4.3







